requirements-test.txt
grpc_tools
mock==4.0.3
black==19.10b0
mypy==0.910
pylint==2.6.0
pytest-cov==2.10.1
pytest-xdist==2.1.0
pytest==6.1.2
ipykernel